Mix Stitcher to fix Vanilla/Forge bug.
This fixes an issue where the stitched texture has a texture size that is
beyond what the client can handle on one dimension (width) but the other
dimension (height) can fit it.

Ex.

Client can handle 8192 and stitched dimensions are 8192 x 4096.

Texture to be stitched is 256x256

Forge says the client should try and stitch downwards before expanding the
width. Unfortunately, the flag isn't reversed and instead of checking if
the height can support adding the texture, it checks the width which is
already at 8192 which is our max. This change flips the incorrect flag and
has it check the height as indeed.

Signed-off-by: Chris Sanders <zidane@outlook.com>