Fix and re-enable recipe duplication detector.
Add ability to specify grabbing a block or item first from the mapper.
This addresses Blocks and Items with the same id and gives differing
priority based on what the node typically would get.

Ex.

SoilNode deals with blocks so we lookup Block first.
BreakNode deals with tools in hand so we lookup Item first.

Signed-off-by: Chris Sanders <zidane@outlook.com>