Mixin cleanup
@Mumfrey , can you look these over and...

A. Tell me why the Inject in TileEntityChestRenderer comes up and says it
can't find func_147500_a.
B. I cannot apply the BlockTallGrass as it says it is being applied twice
(only Mixin that says this and I mentioned this to you in chat).

Thanks for the help ahead of time.

Signed-off-by: Chris Sanders <zidane@outlook.com>