Fix open file ghost handles.
This became a rather large issue for our Mac and Linux users on the Almura
flagship server due to the amount of content we serve. The issue is
summarized as follows:

When java constructs an ImageInputStream, it wraps around an InputStream
and reads data from that internal one. When its done reading (and usually
it creates a BufferedImage), it closes the ImageInputStream but NOT the
internal InputStream.

The closing of the internal InputStream is left up to the finalizer of the
JVM. The finalizer, when an object is about to be garbage collected, will
close the internal InputStream before being ultimately disposed of. IIcon
instances are never disposed of (or, rather, garbage collected).

The simple fix is to just close the stream ourselves when we are sure we
are done reading the stream. On my tests, this lowers the open files from
4k+ added by Almura alone (on the flagship modpack) to 0.

@mcsnetworks .

Signed-off-by: Chris Sanders <zidane@outlook.com>