Change how U/Vs are calculated. Closes #257.
This commit has us return to clipping the atlas before feeding it to the
quad so that the calculations on UVs for the damage model fall in line
with the expectation that your sprite is never an atlas itself.

This change should be forward compatibile; it should work for quads whose
sprite isn't an atlas to begin with but this is untested. Will revise the
code should that end up an issue.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>