Large refactor + Staged Asset Task Execution Loader.
The asset loader is now branched into stages comprised of tasks. The
AssetType defines the tasks that can be ran against it and the stage
determines which tasks it'll run at this time.

The Asset context binds together a Builder, A ConfigurationNode, and the
built Catalog. It is passed through the stage and the stage acts upon the
ConfigurationNode and populates the Builder with characteristcs. The stage
can make the decision to build the catalog object and it isn't required of
stages to only act upon the builder (while not in this commit, the
SetCreativeTabIcon task will act upon the built catalog instead).

This commit only has CreativeTabs loading from tab files as a test to see
if this framework will work. Now that I have the base in, I'll start
writing the tasks to restore our block types.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>