Add Action framework and implement Break.
Since 2.5, Break has received a massive bump in functionality due to new
design changes in the block format. Now its entirely possible to have
multiple "blocks" of break data that target different ItemType usages
and incur different effects. The wiki of this project will be updated
soon with an example.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>