Add debug panel to Origins HUD, Closes #313
The panel struggles when rendering certain blocks (beds, doors,
sugarcane to name a few) but this appears to not be a fault of how I'm
doing things from what I can tell.

UIImage from MalisisCore can equally not render these blocks. We'll have
to find a way to address those that do not render correctly or try and
find a common element between them and not attempt to render the
block if we find one with said element.

Signed-off-by: Steven Downer <grinch@outlook.com>