Create CatalogDelegate system, a placeholder asset struct.
CatalogDelegates simplify the loading process by sparing us of the
age-old "Chicken and Egg" scenario where assets depend on other assets
that depend on those assets (a great example is CreativeTabs need an
ItemStack Tab Icon but Items need a CreativeTab...). With delegates,
we're able to say "The catalog will be this" and lazy-load the actual
Catalog instance later on when either SpongeAPI or the Minecraft engine
requests it.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>