Move to simulation-database-application procedure for list items
In the past we applied list item changes before sending to the database regardless if the database fails. This is ideal for many reasons but the biggest being keeping things in-sync between the engine and the database server. This new approach simulates the modification, applies the changes to the database, and finally applies the changes to the live engine. This commit does not have verification checks against the live engine, those will come later.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>