Do not simulate extractions, play those live.
This is a methodology change I am making as I would rather "refund" items if a database update failed due to a mod bug/etc rather than deal with dupe exploits for the same reasons. I have a TODO to perform the refund but is low priority (as only off cases would have this issue).

Signed-off-by: Chris Sanders <zidane@spongepowered.org>